eliminate the store mixin
