+++ addon/model.js
@@ -5,7 +5,7 @@ import EmberObject, { computed, get, set, defineProperty } from '@ember/object';
 import { isArray } from '@ember/array';
 import { assert, warn } from '@ember/debug';
 import { readOnly } from '@ember/object/computed';
-import { recordDataToRecordMap } from './mixins/store';
+import { recordDataToRecordMap } from './utils/caches';
 
 import { recordDataFor } from './-private';
 import M3RecordArray from './record-array';
