+++ addon/utils/caches.js
@@ -1,0 +1,3 @@
+export const recordDataToRecordMap = new WeakMap();
+export const recordDataToQueryCache = new WeakMap();
+export const recordToRecordArrayMap = new WeakMap();
