+++ addon/utils/resolve.js
@@ -1,5 +1,5 @@
 import { dasherize } from '@ember/string';
-import { recordDataToRecordMap } from '../mixins/store';
+import { recordDataToRecordMap } from '../utils/caches';
 import { CUSTOM_MODEL_CLASS } from 'ember-m3/-infra/features';
 
 export function computeAttributeReference(key, value, modelName, schemaInterface, schema) {
